fix: opt pump cache, issue: #296
