+++ src/backend/booster/bk_dist/handler/ue4/cc/handler.go
@@ -271,7 +271,7 @@ func (cc *TaskCC) analyzeIncludes(dependf string, workdir string) ([]*dcFile.Inf
 		sep = "\r\n"
 	}
 	lines := strings.Split(string(data), sep)
-	uniqlines := commonUtil.UniqArr(lines)
+	uniqlines := dcUtil.UniqArr(lines)
 	blog.Infof("cc: got %d uniq include file from file: %s", len(uniqlines), dependf)
 
 	return dcFile.GetFileInfo(uniqlines, false, false, dcPump.SupportPumpLstatByDir(cc.sandbox.Env))
@@ -357,10 +357,10 @@ func (cc *TaskCC) resolveDependFile(sep, workdir string, includes *[]string) err
 					targetf, _ = filepath.Abs(filepath.Join(workdir, targetf))
 				}
 
-				*includes = append(*includes, commonUtil.FormatFilePath(targetf))
+				*includes = append(*includes, dcUtil.FormatFilePath(targetf))
 
 				// 如果是链接，则将相关指向的文件都包含进来
-				fs := commonUtil.GetAllLinkFiles(targetf)
+				fs := dcUtil.GetAllLinkFiles(targetf)
 				if len(fs) > 0 {
 					*includes = append(*includes, fs...)
 				}
@@ -426,7 +426,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 	// 				targetf, _ = filepath.Abs(filepath.Join(workdir, targetf))
 	// 			}
 
-	// 			includes = append(includes, commonUtil.FormatFilePath(targetf))
+	// 			includes = append(includes, dcUtil.FormatFilePath(targetf))
 	// 		}
 	// 	}
 	// }
@@ -438,7 +438,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 			if !filepath.IsAbs(l) {
 				l, _ = filepath.Abs(filepath.Join(workdir, l))
 			}
-			includes = append(includes, commonUtil.FormatFilePath(l))
+			includes = append(includes, dcUtil.FormatFilePath(l))
 		}
 	}
 
@@ -449,7 +449,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 			if !filepath.IsAbs(l) {
 				l, _ = filepath.Abs(filepath.Join(workdir, l))
 			}
-			includes = append(includes, commonUtil.FormatFilePath(l))
+			includes = append(includes, dcUtil.FormatFilePath(l))
 		}
 	}
 
@@ -460,7 +460,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 			if !filepath.IsAbs(l) {
 				l, _ = filepath.Abs(filepath.Join(workdir, l))
 			}
-			includes = append(includes, commonUtil.FormatFilePath(l))
+			includes = append(includes, dcUtil.FormatFilePath(l))
 		}
 	}
 
@@ -474,7 +474,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 	// for i := range includes {
 	// 	includes[i] = strings.Replace(includes[i], "\\", "/", -1)
 	// }
-	uniqlines := commonUtil.UniqArr(includes)
+	uniqlines := dcUtil.UniqArr(includes)
 
 	// TODO : append symlink or symlinked if need
 	links, _ := getIncludeLinks(cc.sandbox.Env, uniqlines)
@@ -483,7 +483,7 @@ func (cc *TaskCC) copyPumpHeadFile(workdir string) error {
 	}
 
 	// TODO :将链接路径找出并放到前面
-	linkdirs := commonUtil.GetAllLinkDir(uniqlines)
+	linkdirs := dcUtil.GetAllLinkDir(uniqlines)
 	if len(linkdirs) > 0 {
 		uniqlines = append(linkdirs, uniqlines...)
 	}
